rpi_flow/requirements.txt
numpy
opencv-python
pyserial
smbus2
# Optional (recommended on Raspberry Pi OS with libcamera)
picamera2

